Fix up some printf() type warnings

+++ libopendkim/tests/t-signperf.c
@@ -286,7 +286,7 @@ main(int argc, char **argv)
 	fprintf(stdout,
 	        "*** SIGNING SPEED TEST: %s/%s with %s, size %u for %lds\n",
 	        canon_name(hcanon), canon_name(bcanon), alg_name(signalg),
-	        msgsize, (long) testint);
+	        (unsigned int) msgsize, (long) testint);
 
 	key = KEY;
